refactor, move some function to TypeUtil class
